requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
passlib[bcrypt]
python-dotenv
pydantic[email]
python-jose[cryptography]
python-multipart
jinja2
pytest
requests
pytest-order
pytest-html